requirements.txt
langchain
openai
python-dotenv
edge-tts
SpeechRecognition
pyaudio
openai-whisper
numpy
sounddevice
pygame